pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "net-topology"
version = "0.1.0"
description = "Network topology scanner — discover devices, collect inventory, map links"
requires-python = ">=3.10"
dependencies = [
    "pysnmp>=4.4,<6.0",
    "routeros-api>=0.17,<1.0",
    "pyyaml>=6.0,<7.0",
]

[project.scripts]
net-topology = "net_topology.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
